# apps/backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Auth API
  description: API for user authentication, including sign-up and sign-in.
  version: 1.0.0
security:
  - bearerAuth: []
paths:
  /api/v1/auth/sign-up:
    x-tag: Authentication
    post:
      summary: User Sign-Up
      description: Creates a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: securepassword
                name:
                  type: string
                  example: John Doe
              required:
                - email
                - password
                - name
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

  /api/v1/auth/sign-in:
    x-tag: Authentication
    post:
      summary: User Sign-In
      description: Authenticates a user and returns a token in an HTTP-only cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: securepassword
              required:
                - email
                - password
      responses:
        "200":
          description: User authenticated successfully.
          headers:
            Set-Cookie:
              description: HTTP-only cookie containing the JWT token.
              schema:
                type: string
                example: token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Max-Age=3600
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid credentials
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

  /api/v1/auth/me:
    x-tag: Authentication
    get:
      summary: Get authenticated user
      description: Retrieves the authenticated user's details.
      security:
        - bearerAuth: []

      responses:
        "200":
          description: User details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: User ID
                  email:
                    type: string
                    description: User email
                  name:
                    type: string
                    description: User name
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized

  /api/v1/products:
    x-tag: Products
    get:
      summary: Get all products
      description: Retrieves a list of all products.
      responses:
        "200":
          description: Products retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: UUID of the product
                        name:
                          type: string
                          description: Name of the product
                        categoryId:
                          type: string
                          description: UUID of the category
        "500":
          description: Internal server error.
    post:
      summary: Create a new product
      description: Creates a new product with the provided name and category ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the product
                categoryId:
                  type: string
                  description: UUID of the category
              required:
                - name
                - categoryId
      responses:
        "201":
          description: Product created successfully.
        "400":
          description: Validation error.
        "500":
          description: Internal server error.

  /api/v1/products/{id}:
    x-tag: Products
    get:
      summary: Get product by ID
      description: Retrieves a product by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the product
      responses:
        "200":
          description: Product retrieved successfully.
        "404":
          description: Product not found.
        "500":
          description: Internal server error.

    put:
      summary: Update a product
      description: Updates a product's details by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Updated name of the product
                categoryId:
                  type: string
                  description: Updated category ID of the product
      responses:
        "200":
          description: Product updated successfully.
        "400":
          description: Validation error.
        "500":
          description: Internal server error.

    delete:
      summary: Delete a product
      description: Deletes a product by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the product
      responses:
        "200":
          description: Product deleted successfully.
        "500":
          description: Internal server error.

  /api/v1/products/{id}/feedback:
    x-tag: Products
    get:
      summary: Get feedback for a product
      description: Retrieves feedback for a specific product by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the product
      responses:
        "200":
          description: Feedback retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: Feedback ID
                    comment:
                      type: string
                      description: Feedback comment
                    rating:
                      type: integer
                      description: Feedback rating
                    userId:
                      type: string
                      description: ID of the user who submitted the feedback
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Product not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

  /api/v1/categories:
    x-tag: Categories
    get:
      summary: Get all categories
      description: Retrieves a list of all categories.
      responses:
        "200":
          description: Categories retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: UUID of the category
                        name:
                          type: string
                          description: Name of the category
        "500":
          description: Internal server error.

    post:
      summary: Create a new category
      description: Creates a new category with the provided name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the category
              required:
                - name
      responses:
        "201":
          description: Category created successfully.
        "400":
          description: Validation error.
        "500":
          description: Internal server error.

  /api/v1/categories/{id}:
    x-tag: Categories
    get:
      summary: Get category by ID
      description: Retrieves a category by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the category
      responses:
        "200":
          description: Category retrieved successfully.
        "404":
          description: Category not found.
        "500":
          description: Internal server error.

    put:
      summary: Update a category
      description: Updates a category's details by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Updated name of the category
                description:
                  type: string
                  description: Updated description of the category
      responses:
        "200":
          description: Category updated successfully.
        "400":
          description: Validation error.
        "500":
          description: Internal server error.

    delete:
      summary: Delete a category
      description: Deletes a category by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the category
      responses:
        "200":
          description: Category deleted successfully.
        "500":
          description: Internal server error.

  /api/v1/feedback:
    x-tag: Feedback
    post:
      summary: Submit feedback
      description: Allows users to submit feedback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: ID of the user submitting the feedback
                message:
                  type: string
                  description: Feedback message
                rating:
                  type: integer
                  description: Rating provided by the user (1-5)
                  minimum: 1
                  maximum: 5
              required:
                - userId
                - message
                - rating
      responses:
        "201":
          description: Feedback submitted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: ID of the created feedback
                      userId:
                        type: string
                        description: ID of the user who submitted the feedback
                      message:
                        type: string
                        description: Feedback message
                      rating:
                        type: integer
                        description: Rating provided by the user
                      createdAt:
                        type: string
                        format: date-time
                        description: Timestamp of feedback creation
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 400
                  error:
                    type: string
                    example: Validation error
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal server error

    get:
      summary: Get all feedback
      description: Retrieves a list of all feedback entries.
      responses:
        "200":
          description: Feedback retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: UUID of the feedback
                        userId:
                          type: string
                          description: UUID of the user who submitted the feedback
                        message:
                          type: string
                          description: Feedback message
                        rating:
                          type: integer
                          description: Rating provided by the user
                        createdAt:
                          type: string
                          format: date-time
                          description: Timestamp of feedback creation
        "500":
          description: Internal server error.

  /api/v1/feedback/{id}:
    x-tag: Feedback
    get:
      summary: Get feedback by ID
      description: Retrieves feedback details by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the feedback
      responses:
        "200":
          description: Feedback retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: ID of the feedback
                      userId:
                        type: string
                        description: ID of the user who submitted the feedback
                      message:
                        type: string
                        description: Feedback message
                      rating:
                        type: integer
                        description: Rating provided by the user
                      createdAt:
                        type: string
                        format: date-time
                        description: Timestamp of feedback creation
        "404":
          description: Feedback not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  error:
                    type: string
                    example: Feedback not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal server error

    put:
      summary: Update feedback
      description: Updates feedback details by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: Updated feedback message
                rating:
                  type: integer
                  description: Updated rating (1-5)
              required:
                - message
      responses:
        "200":
          description: Feedback updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: Feedback updated successfully
        "400":
          description: Validation error.
        "500":
          description: Internal server error.

    delete:
      summary: Delete feedback
      description: Deletes feedback by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the feedback
      responses:
        "200":
          description: Feedback deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: Feedback deleted successfully
        "500":
          description: Internal server error.

  /api/v1/users/{id}:
    x-tag: Users
    get:
      summary: Get user by ID
      description: Retrieves a user by their ID.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the user
      responses:
        "200":
          description: User retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: User ID
                  email:
                    type: string
                    description: User email
                  name:
                    type: string
                    description: User name
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error